⎆

+++ frontend/src/components/SettingsGoogleDrive.tsx
@@ -1,12 +1,4 @@
-// TODO
-
-// https://developers.google.com/identity/protocols/oauth2/web-server#node.js_1
-//
-// Access tokens expire. This library will automatically use a refresh token to
-// obtain a new access token if it is about to expire. An easy way to make sure
-// you always store the most recent tokens is to use the tokens event:
-
-import _ from "lodash";
+import _, { set } from "lodash";
 import { useEffect, useState } from "react";
 import { Link as RouterLink, useNavigate } from "react-router-dom";
 import useSWR from "swr";
@@ -15,6 +7,11 @@ import {
   Button,
   Checkbox,
   CircularProgress,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogContentText,
+  DialogTitle,
   Fade,
   FormControl,
   FormControlLabel,
@@ -32,6 +29,7 @@ import supabase, { useAuth } from "../supabase";
 import useErrorBar from "../hooks/useErrorBar";
 import { DefaultService } from "../client";
 import useGoogleDrive from "../hooks/useGoogleDrive";
+import { useAsyncEffect } from "../hooks/useAsyncEffect";
 
 interface File {
   id: string;
@@ -46,6 +44,12 @@ export default function SettingsGoogleDrive() {
 
   const [files, setFiles] = useState<File[] | null>(null);
   const { showError } = useErrorBar();
+
+  // dialog
+  const [checkDeleteId, setCheckDeleteId] = useState<string | null>(null);
+  const [confirmedDeleteId, setConfirmedDeleteId] = useState<string | null>(
+    null
+  );
 
   // check that we are logged in
   useEffect(() => {
@@ -97,18 +101,34 @@ export default function SettingsGoogleDrive() {
         throw Error("Could not start sync job");
       }
     } else {
-      const { error } = await supabase
-        .from("synced_folder")
-        .delete()
-        .match({ user_id: session!.user.id, remote_id: fileId });
-      if (error) {
-        console.error(error);
-        showError();
-        throw Error("Could not delete synced folder");
-      }
-      mutate(_.reject(syncedFolders, { remote_id: fileId }), false);
+      // unchecked. use a dialog to confirm
+      setCheckDeleteId(fileId);
     }
   };
+
+  useAsyncEffect(
+    async () => {
+      if (confirmedDeleteId !== null) {
+        setCheckDeleteId(null);
+        setConfirmedDeleteId(null);
+        const { error } = await supabase
+          .from("synced_folder")
+          .delete()
+          .match({ user_id: session!.user.id, remote_id: confirmedDeleteId });
+        if (error) {
+          console.error(error);
+          showError();
+          throw Error("Could not delete synced folder");
+        }
+        mutate(
+          _.reject(syncedFolders, { remote_id: confirmedDeleteId }),
+          false
+        );
+      }
+    },
+    async () => {},
+    [confirmedDeleteId]
+  );
 
   // On access token change, check for drive access
   useEffect(() => {
@@ -137,87 +157,128 @@ export default function SettingsGoogleDrive() {
   }, [google.gapi]);
 
   return (
-    <Container>
-      <Typography variant="h4">Configure Google Drive</Typography>
-      <List>
-        <ListItem>
-          <Button
-            onClick={google.signIn}
-            disabled={google.isLoading || google.accessToken !== null}
-          >
-            Connect Google Drive
-          </Button>
-          <Button
-            onClick={async () => {
-              setFiles(null);
-              await google.signOut();
-            }}
-            disabled={google.isLoading || google.accessToken === null}
-          >
-            Disconnect Google Drive
-          </Button>
-        </ListItem>
-        {!google.isLoading && google.accessToken !== null && (
+    <>
+      <Container>
+        <Typography variant="h4">Configure Google Drive</Typography>
+        <List>
           <ListItem>
-            <Button to="/files" component={RouterLink}>
-              Go to Files
-            </Button>{" "}
-            to see your synced files
-          </ListItem>
-        )}
-        <FormControl
-          disabled
-          size="small"
-          sx={{
-            display: "flex",
-            flex: "0 5 auto",
-            width: "200px",
-          }}
-        >
-          <InputLabel>Project</InputLabel>
-          <Select label="Project" value="default" autoWidth>
-            <MenuItem value="default">Default</MenuItem>
-          </Select>
-        </FormControl>
-        {files !== null && (
-          <>
-            <Typography variant="h6">Folders:</Typography>
-            <FormGroup>
-              {files.map((file, i) => (
-                <FormControlLabel
-                  key={i}
-                  label={file.name}
-                  control={<Checkbox />}
-                  checked={syncedFolders?.some(
-                    (syncedFolder) => syncedFolder.remote_id === file.id
-                  )}
-                  onChange={(_, checked) =>
-                    onUpdateSyncedFolders(checked, file.id)
-                  }
-                />
-              ))}
-              {files.length === 100 && (
-                <Typography variant="body2">
-                  Only the first 100 folders are shown
-                </Typography>
-              )}
-            </FormGroup>
-          </>
-        )}
-        <ListItem>
-          {google.isLoading && (
-            <Fade
-              in={google.isLoading}
-              style={{
-                transitionDelay: google.isLoading ? "800ms" : "0ms",
-              }}
-              unmountOnExit
+            <Button
+              onClick={google.signIn}
+              disabled={google.isLoading || google.accessToken !== null}
             >
-              <CircularProgress />
-            </Fade>
+              Connect Google Drive
+            </Button>
+            <Button
+              onClick={async () => {
+                setFiles(null);
+                await google.signOut();
+              }}
+              disabled={google.isLoading || google.accessToken === null}
+            >
+              Disconnect Google Drive
+            </Button>
+          </ListItem>
+          {!google.isLoading && google.accessToken !== null && (
+            <ListItem>
+              <Button to="/files" component={RouterLink}>
+                Go to Files
+              </Button>{" "}
+              to see your synced files
+            </ListItem>
           )}
-        </ListItem>
-      </List>
-    </Container>
+          <FormControl
+            disabled
+            size="small"
+            sx={{
+              display: "flex",
+              flex: "0 5 auto",
+              width: "200px",
+            }}
+          >
+            <InputLabel>Project</InputLabel>
+            <Select label="Project" value="default" autoWidth>
+              <MenuItem value="default">Default</MenuItem>
+            </Select>
+          </FormControl>
+          {files !== null && (
+            <>
+              <Typography variant="h6">Folders:</Typography>
+              <FormGroup>
+                {files.map((file, i) => (
+                  <FormControlLabel
+                    key={i}
+                    label={file.name}
+                    control={<Checkbox />}
+                    checked={syncedFolders?.some(
+                      (syncedFolder) => syncedFolder.remote_id === file.id
+                    )}
+                    onChange={(_, checked) =>
+                      onUpdateSyncedFolders(checked, file.id)
+                    }
+                  />
+                ))}
+                {files.length === 100 && (
+                  <Typography variant="body2">
+                    Only the first 100 folders are shown
+                  </Typography>
+                )}
+              </FormGroup>
+            </>
+          )}
+          <ListItem>
+            {google.isLoading && (
+              <Fade
+                in={google.isLoading}
+                style={{
+                  transitionDelay: google.isLoading ? "800ms" : "0ms",
+                }}
+                unmountOnExit
+              >
+                <CircularProgress />
+              </Fade>
+            )}
+          </ListItem>
+        </List>
+      </Container>
+
+      {/* Dialogs */}
+      <Dialog
+        open={checkDeleteId !== null}
+        onClose={() => {
+          setCheckDeleteId(null);
+          setConfirmedDeleteId(null);
+        }}
+        aria-labelledby="alert-dialog-title"
+        aria-describedby="alert-dialog-description"
+      >
+        <DialogTitle id="alert-dialog-title">
+          {"Stop syncing folder?"}
+        </DialogTitle>
+        <DialogContent>
+          <DialogContentText id="alert-dialog-description">
+            Are you sure you want to stop syncing this folder? Your annotations
+            will be deleted.
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button
+            onClick={() => {
+              setConfirmedDeleteId(null);
+              setCheckDeleteId(null);
+            }}
+          >
+            No, keep keep syncing
+          </Button>
+          <Button
+            onClick={() => {
+              setConfirmedDeleteId(checkDeleteId);
+              setCheckDeleteId(null);
+            }}
+          >
+            Yes, stop syncing
+          </Button>
+        </DialogActions>
+      </Dialog>
+    </>
   );
 }
